Output all differences between the compared derivations
Before this change, if the set of input names to a derivation changed,
then the set of matching-names inputs was not compared.

This fixes the issue, allowing to use `nix-diff` as a tool for system
administration: there, you have to know all the differences between the
two system derivations, not only until an input set has a mismatching
name (eg. a linux-kernel upgrade should not hide an nginx upgrade).

Fixes #5.