Accept realised store paths as input
The tool still accepts `.drv` files as input but also realised (built)
paths.

It's possible to ask for the derivation corresponding to a store path
with

    nix-store -qd store-path

Nix also takes care of following symbolic links.

The motivation is being able to easily compare a newly built system with
the currently running system:

    nix-diff /run/current-system $result