[#58] Add intermediate representation
Problem: There is no way to reuse result of diff alghorithm because
program print it into stdout in human readable format in-place.

Solution: divide steps of finding diff between two derivations and
printing it out. Now result of `diff` is tree with result of derivation
comparison, that you can print than in stdout as was before, or use in
your program.