Support relocated store; lightweight store module (#82)
Support relocated store when configured with the `NIX_STORE_DIR` and
`NIX_REMOTE` environment variables.

This allows `nix-diff` to be used in the sandbox with instantiations in
a store whose physical location is in the build directory rather than
the logical `storeDir` (`/nix/store`).

This is achieved by changing introducing a `newtype StorePath` to help
distinguish paths that may be logical paths rather than physical ones.
It is not a true abstraction in the sense that Nix's own store concept
is, but that's ok. It makes this use case work, while still only relying
on the ambient environment, requiring no types and variables for the
store itself.

I plan to use this functionality for error reporting in evaluation tests
that evaluate Nixpkgs for the purpose of testing Nix itself, in the
context of Nixpkgs "passthru" tests and Nix's own CI setup.